requirements.txt
Django~=4.1.0
python-dotenv~=0.21.1
Pillow==9.2.0
celery==5.2.7
flower==1.1.0
stripe==4.0.2
WeasyPrint==56.1
redis==4.3.4
django-rosetta==0.9.8
django-parler==2.3
django-localflavor==3.1